# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'Planora'
  description: 'Web Resources Specification (A7) for Planora'
servers:
  - url: http://localhost:8000
    description: Localhost server
...
paths:
  /login:
    get:
      operationId: R101
      summary: 'Show Login Form'
      description: 'Returns the login form for users to authenticate. PUB'
      tags:
        - 'M01: Authentication and Profile'
      responses:
        '200':
          description: 'Successful response. Displays the login form.'
    post:
      operationId: R102
      summary: 'Authenticate User'
      description: 'Processes the login form submission and authenticates the user. Redirects on success or failure. PUB'
      tags:
        - 'M01: Authentication and Profile'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password

      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          headers:
            Location:
              schema:
                type: string
                example:
                  302Success:
                    description: 'Successful authentication. Redirect to user dashboard.'
                    value: '/projects'
                  302Error:
                    description: 'Failed authentication. Redirect to login form.'
                    value: '/login'
...
  /tasks/{id}:
    get:
      operationId: R301
      summary: 'View Task'
      description: 'Displays the details of a specific task, including its status. Access: ADM, OWN, PRM, PRC'
      tags:
        - 'M03: Tasks'
      parameters:
        - name: id
          in: path
          required: true
          description: 'The ID of the task to view.'
          schema:
            type: integer
      responses:
        '200':
          description: 'Task details displayed successfully.'
        '500':
          description: 'Failed to show task.'
    put:
      operationId: R303
      summary: 'Update Task'
      description: 'Updates the details of an existing task. Access: ADM, OWN, PRM, PRC'
      tags:
        - 'M03: Tasks'
      parameters:
        - name: id
          in: path
          required: true
          description: 'The ID of the task to update.'
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                task_title:
                  type: string
                task_description:
                  type: string
                deadline:
                  type: string
                  format: date
                task_status_id:
                  type: integer
                  description: 'The status ID of the task.'
                task_priority:
                  type: string
                  enum: [ Low, Medium, High ]
              required:
                - task_title
                - task_status_id
      responses:
        '200':
          description: 'Task updated successfully.'
        '500':
          description: 'Failed to update task. Please try again.'
    delete:
      operationId: R304
      summary: 'Delete Task'
      description: 'Deletes a specific task. Access: ADM, OWN, PRM, PRC'
      tags:
        - 'M03: Tasks'
      parameters:
        - name: id
          in: path
          required: true
          description: 'The ID of the task to delete.'
          schema:
            type: integer
      responses:
        '200':
          description: 'Task deleted successfully.'
        '500':
          description: 'Failed to delete task. Please try again.'
...
  /admin:
    get:
      operationId: R501
      summary: 'Search users'
      description: 'Search for users by username and email. Access: ADM'
      tags:
        - 'M05: Admin'
      parameters:
        - name: project_id
          in: path
          required: false
          description: 'The id of the project to search members.'
          schema:
            type: integer
        - name: user
          in: path
          required: true
          description: 'The user to search for.'
          schema:
            type: string
      responses:
        '200':
          description: 'Search results displayed successfully.'
        '500':
          description: 'Failed to load user list.'